+++ gostringsr2/__init__.py
@@ -4,6 +4,6 @@
 
 __author__ = """Jonathan Wrightsell"""
 __email__ = "jonathan.wrightsell@carvesystems.com"
-__version__ = "1.1.1"
+__version__ = "1.1.2"
 
 from .gostringsr2 import *